requirements.txt
tensorflow==2.17.0
tensorflow-probability==0.24.0
tf-keras==2.17.0
pandas==2.2.2
pymongo==4.8.0
python-dotenv==1.0.1
flask==3.0.3
matplotlib==3.9.2
flask-cors==5.0.0
redis==5.0.8
rq==1.16.2
requests==2.32.3